[feat]: 마이페이지지의 내가 단 리뷰목록과 좋아요 표시한 리뷰 목록 조회 기능 구현

+++ app/src/main/java/com/example/movieismylife/model/CommentView.kt
@@ -3,7 +3,7 @@ package com.example.movieismylife.model
 import com.google.firebase.Timestamp
 
 data class CommentView (
-    val score: Long = 0,
+    val score: Float = 0f,
     val content: String = "",
     val name: String = "",
     val profile: String = "",
